Style register page
