+++ client/src/pages/Register/Register.module.scss
@@ -1,0 +1,135 @@
+.container {
+  --background-color: var(--primary-background);
+  --text-color-dark: var(--primary-text);
+  --text-color-light: var(--secondary-text);
+  --link-color: var(--primary-button);
+  --border-color: var(--separator);
+
+  margin-top: 10px;
+  text-align: center;
+  font-size: 14px;
+  max-width: 350px;
+  // padding: 10px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 10px;
+
+  .logo {
+    margin: 50px 0 10px 0;
+  }
+
+  .formContainer {
+    border: 1px solid var(--border-color);
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 10px;
+
+    .form {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      padding: 0 40px 40px 40px;
+      gap: 6px;
+
+      .mainText {
+        color: var(--text-color-light);
+        font-weight: 700;
+        font-size: 16px;
+      }
+
+      .facebookBtn {
+        width: 100%;
+        justify-content: center;
+        margin: 10px 0;
+        height: 34px;
+      }
+
+      .or {
+        width: 100%;
+        position: relative;
+        height: 38px;
+        margin-bottom: 6px;
+
+        .line {
+          position: absolute;
+          top: 50%;
+          transform: translateY(-50%);
+          height: 1px;
+          width: 100%;
+          background: var(--border-color);
+        }
+
+        .orText {
+          position: absolute;
+          left: 50%;
+          top: 50%;
+          transform: translate(-50%, -50%);
+          text-transform: uppercase;
+          color: var(--text-color-light);
+          background: var(--background-color);
+          padding: 0 10px;
+          font-weight: 700;
+          font-size: 14px;
+        }
+      }
+
+      .formControl {
+        position: relative;
+        width: 100%;
+
+        .input {
+          border: 1px solid var(--border-color);
+          color: var(--text-color-dark);
+          width: 100%;
+          height: 38px;
+          padding: 8px;
+          border-radius: 3px;
+
+          &:focus {
+            outline: none;
+            border-color: var(--text-color-light);
+          }
+        }
+
+        .placeholder {
+          color: var(--text-color-light);
+          font-size: 12px;
+          position: absolute;
+          top: 50%;
+          transform: translateY(-50%);
+          left: 8px;
+        }
+
+        .label {
+          color: var(--text-color-light);
+          font-size: 10px;
+          position: absolute;
+          left: 8px;
+          top: 3px;
+          transition: 0.2s ease all;
+        }
+      }
+
+      .signupBtn {
+        width: 100%;
+        margin-top: 20px;
+      }
+    }
+  }
+
+  .loginPrompt {
+    border: 1px solid var(--border-color);
+    width: 100%;
+    height: 70px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    .login {
+      text-decoration: none;
+      color: var(--link-color);
+    }
+  }
+}
